Add test runs for macOS, Windows and rdkafka extension compatibility to travis (#8)
* Add macOS to travis
* Add windows to travis
* Add rdkafka ext test to travis and improve windows
* Mark CallbackProxyTest as ffiOnly
* Use inline CallbackProxy stub in tests
** allow override of constructor in subclasses
** mainly to run tests against rdkafka ext (CallbackProxy unknown)
* Support ext in librdkafka version detection in tests
* Fix librdkafka version detection in tests
* Mark test KafkaConsumerTest::commitWithOffsetAndMetadata ffiOnly
* Simplify MetaData topic name test
* Remove not supported functions on windows from cdef
* Disable class binding generation
* Add tests timeouts and customize admin client result event timeout
* rise create & delete topic timeouts
* move test config from phpunit xml to config.php file

+++ tests/RdKafka/ConfTest.php
@@ -220,7 +220,7 @@ class ConfTest extends TestCase
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(0, 0, __METHOD__ . '1');
         $producerTopic->produce(0, 0, __METHOD__ . '2');
-        $producer->poll((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->poll(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertCount(2, $drMsgCallbackStack);
         $this->assertSame($producer, $drMsgCallbackStack[0]['producer']);
@@ -228,7 +228,7 @@ class ConfTest extends TestCase
         $this->assertSame($producer, $drMsgCallbackStack[1]['producer']);
         $this->assertSame(__METHOD__ . '2', $drMsgCallbackStack[1]['message']->payload);
 
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
     }
 
     public function testSetStatsCb(): void
