+++ tests/RdKafka/ProducerTopicTest.php
@@ -55,7 +55,7 @@ class ProducerTopicTest extends TestCase
 
         $topic->produce(RD_KAFKA_PARTITION_UA, 0, __METHOD__, 'key-topic-produce');
 
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertSame(__METHOD__, $payload);
     }
@@ -98,7 +98,7 @@ class ProducerTopicTest extends TestCase
 
         $topic->producev(RD_KAFKA_PARTITION_UA, 0, null, 'key-topic-produce');
 
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertSame(null, $payload);
     }
@@ -126,7 +126,7 @@ class ProducerTopicTest extends TestCase
             ['header-name-topic-produce' => 'header-value-topic-produce']
         );
 
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertSame(__METHOD__, $payload);
         $this->assertSame(['header-name-topic-produce' => 'header-value-topic-produce'], $headers);
@@ -156,7 +156,7 @@ class ProducerTopicTest extends TestCase
             123456789
         );
 
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertSame(__METHOD__, $payload);
         $this->assertSame(123456789, $timestamp);
